add component index exports and refactor import paths
