+++ resources/js/components/screens/categories/index.ts
@@ -1,0 +1,7 @@
+export { default as Actions } from './Actions.vue';
+export { default as Delete } from './Delete.vue';
+export { default as Form } from './Form.vue';
+export { default as Item } from './Item.vue';
+export { default as List } from './List.vue';
+export { default as Stat } from './Stat.vue';
+export { default as Stats } from './Stats.vue';
