+++ resources/js/components/screens/settings/AppearanceTabs.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { useAppearance } from '@/composables/useAppearance';
+import { useAppearance } from '@/composables/use-appearance';
 
 import { Monitor, Moon, Sun } from 'lucide-vue-next';
