+++ resources/js/components/screens/settings/index.ts
@@ -1,0 +1,9 @@
+export { default as AppearanceTabs } from './AppearanceTabs.vue';
+export { default as AvatarUpload } from './AvatarUpload.vue';
+export { default as DangerZone } from './DangerZone.vue';
+export { default as DataExport } from './DataExport.vue';
+export { default as InputPassword } from './InputPassword.vue';
+export { default as InputSelect } from './InputSelect.vue';
+export { default as InputString } from './InputString.vue';
+export { default as InputToggle } from './InputToggle.vue';
+export { default as SubmitButton } from './SubmitButton.vue';
