+++ resources/js/pages/auth/Login.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { TextLink } from '@/components/elements';
 import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
 import { Error, Input } from '@/components/ui/input';
@@ -6,8 +7,6 @@ import { Label } from '@/components/ui/label';
 import { AuthLayout } from '@/layouts/auth-layout';
 import { Form } from '@inertiajs/vue3';
 import { LoaderCircle } from 'lucide-vue-next';
-
-import TextLink from '@/components/elements/TextLink.vue';
 
 defineProps<{
     status?: string;
