+++ resources/js/pages/auth/Register.vue
@@ -1,12 +1,11 @@
 <script setup lang="ts">
+import { TextLink } from '@/components/elements';
 import { Button } from '@/components/ui/button';
 import { Error, Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { AuthLayout } from '@/layouts/auth-layout';
 import { Form } from '@inertiajs/vue3';
 import { LoaderCircle } from 'lucide-vue-next';
-
-import TextLink from '@/components/elements/TextLink.vue';
 </script>
 
 <template>
